+++ src/main/scala-2.11/markets/ExchangeActor.scala
@@ -18,7 +18,6 @@ package markets
 import akka.actor.{ActorRef, Props}
 import akka.agent.Agent
 
-import markets.clearing.engines.MatchingEngine
 import markets.orders.Order
 import markets.tickers.Tick
 import markets.tradables.Tradable
