+++ src/test/scala-2.11/markets/MarketActorSpec.scala
@@ -20,8 +20,7 @@ import akka.agent.Agent
 import akka.testkit.{TestActorRef, TestKit, TestProbe}
 
 import java.util.UUID
-
-import markets.clearing.engines.BrokenMatchingEngine
+import markets.engines.BrokenMatchingEngine
 import markets.orders.limit.{LimitAskOrder, LimitBidOrder}
 import markets.orders.market.{MarketAskOrder, MarketBidOrder}
 import markets.tickers.Tick
