+++ src/test/scala-2.11/markets/engines/BrokenMatchingEngine.scala
@@ -13,8 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.clearing.engines
+package markets.engines
 
+import markets.clearing.engines.{Matching, MutableMatchingEngine}
 import markets.orders.{BidOrder, AskOrder, Order}
 
 import scala.collection.{immutable, mutable}
