+++ src/test/scala-2.11/markets/engines/CDAMatchingEngineSpec.scala
@@ -13,13 +13,14 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.clearing.engines
+package markets.engines
 
 import akka.actor.{ActorRef, ActorSystem}
 import akka.testkit.TestKit
 
 import java.util.UUID
 
+import markets.clearing.engines.Matching
 import markets.orders.limit.{LimitAskOrder, LimitBidOrder}
 import markets.orders.market.{MarketAskOrder, MarketBidOrder}
 import markets.orders.orderings.ask.AskPriceTimeOrdering
